% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_gt_txt.R
\name{correct_gt_txt}
\alias{correct_gt_txt}
\title{Launches the interactive addin to correct "ground truth" .gt.txt files in a folder.}
\usage{
correct_gt_txt()
}
\description{
Launches the interactive addin to correct "ground truth" .gt.txt files in a folder.
}
